encapsulated non-error status codes and added a custom enum for error codes

+++ src/ProblemResults/ResultFactory.cs
@@ -19,7 +19,7 @@ public static class ResultFactory
         {
             Type = ProblemDefaults.Type(problem.Status),
             Title = ProblemDefaults.Title(problem.Status),
-            Status = problem.Status,
+            Status = (int)problem.Status,
             Detail = problem.Detail,
             Instance = $"{httpContext.Request.Method} {httpContext.Request.Path}"
         };
